+++ Language/KURE/Combinators/Transform.hs
@@ -1,4 +1,6 @@
-{-# Language CPP, InstanceSigs #-}
+{-# LANGUAGE CPP #-}
+{-# LANGUAGE DeriveDataTypeable #-}
+{-# LANGUAGE InstanceSigs #-}
 -- |
 -- Module: Language.KURE.Combinators.Transform
 -- Copyright: (c) 2012--2014 The University of Kansas
@@ -58,6 +60,9 @@ import Control.Monad (liftM,ap)
 
 import Data.Foldable
 import Data.Traversable
+#if __GLASGOW_HASKELL__ >= 708
+import Data.Typeable
+#endif
 
 import Language.KURE.Combinators.Arrow
 import Language.KURE.Combinators.Monad
@@ -205,6 +210,9 @@ checkSuccessPBool msg m = do PBool b a <- m
 --   causes a sequence of rewrites to succeed if at least one succeeds, converting failures to
 --   identity rewrites.
 newtype AnyR m a = AnyR (m (PBool a))
+#if __GLASGOW_HASKELL__ >= 708
+  deriving Typeable
+#endif
 
 unAnyR :: AnyR m a -> m (PBool a)
 unAnyR (AnyR mba) = mba
@@ -264,6 +272,9 @@ unwrapAnyR = resultT (checkSuccessPBool "anyR failed" . unAnyR)
 -- | The 'OneR' transformer, in combination with 'wrapOneR' and 'unwrapOneR',
 --   causes a sequence of rewrites to only apply the first success, converting the remainder (and failures) to identity rewrites.
 newtype OneR m a = OneR (Bool -> m (PBool a))
+#if __GLASGOW_HASKELL__ >= 708
+  deriving Typeable
+#endif
 
 unOneR :: OneR m a -> Bool -> m (PBool a)
 unOneR (OneR mba) = mba