+++ Language/KURE/ExtendableContext.hs
@@ -1,4 +1,8 @@
-{-# LANGUAGE InstanceSigs, MultiParamTypeClasses, FlexibleInstances, UndecidableInstances #-}
+{-# LANGUAGE DeriveDataTypeable #-}
+{-# LANGUAGE FlexibleInstances #-}
+{-# LANGUAGE InstanceSigs #-}
+{-# LANGUAGE MultiParamTypeClasses #-}
+{-# LANGUAGE UndecidableInstances #-}
 -- |
 -- Module: Language.KURE.ExtendableContext
 -- Copyright: (c) 2012--2014 The University of Kansas
@@ -21,6 +25,8 @@ module Language.KURE.ExtendableContext
         , extraContext
 ) where
 
+import Data.Typeable
+
 import Language.KURE.Path
 
 ------------------------------------------------------------------------------------------------
@@ -32,6 +38,7 @@ data ExtendContext c e = ExtendContext
                              -- | Retrieve the extra contextual information.
                            , extraContext  :: e
                            }
+  deriving Typeable
 
 -- | Extend a context with some additional information.
 extendContext :: e -> c -> ExtendContext c e