+++ Language/KURE/Injection.hs
@@ -1,4 +1,9 @@
-{-# LANGUAGE InstanceSigs, MultiParamTypeClasses, FlexibleInstances #-}
+{-# LANGUAGE CPP #-}
+{-# LANGUAGE DeriveDataTypeable #-}
+{-# LANGUAGE FlexibleInstances #-}
+{-# LANGUAGE InstanceSigs #-}
+{-# LANGUAGE MultiParamTypeClasses #-}
+{-# LANGUAGE StandaloneDeriving #-}
 -- |
 -- Module: Language.KURE.Injection
 -- Copyright: (c) 2012--2014 The University of Kansas
@@ -34,6 +39,10 @@ module Language.KURE.Injection
 
 import Control.Arrow
 
+#if __GLASGOW_HASKELL__ >= 708
+import Data.Typeable
+#endif
+
 import Language.KURE.Transform
 
 -------------------------------------------------------------------------------
@@ -46,6 +55,10 @@ import Language.KURE.Transform
 class Injection a u where
   inject  :: a -> u
   project :: u -> Maybe a
+
+#if __GLASGOW_HASKELL__ >= 708
+deriving instance Typeable Injection
+#endif
 
 -- | There is an identity injection for all types.
 instance Injection a a where