+++ Language/KURE/Path.hs
@@ -1,4 +1,10 @@
-{-# LANGUAGE CPP, InstanceSigs, MultiParamTypeClasses, FunctionalDependencies, FlexibleInstances #-}
+{-# LANGUAGE CPP #-}
+{-# LANGUAGE DeriveDataTypeable #-}
+{-# LANGUAGE FlexibleInstances #-}
+{-# LANGUAGE FunctionalDependencies #-}
+{-# LANGUAGE InstanceSigs #-}
+{-# LANGUAGE MultiParamTypeClasses #-}
+{-# LANGUAGE StandaloneDeriving #-}
 -- |
 -- Module: Language.KURE.Path
 -- Copyright: (c) 2012--2014 The University of Kansas
@@ -43,6 +49,8 @@ import Data.Monoid
 
 import Control.Arrow ((>>^))
 
+import Data.Typeable
+
 import Language.KURE.Transform
 import Language.KURE.Combinators.Transform
 import Language.KURE.Injection
@@ -56,7 +64,7 @@ type Path crumb = [crumb]
 -------------------------------------------------------------------------------
 
 -- | A 'SnocPath' is a list stored in reverse order.
-newtype SnocPath crumb = SnocPath [crumb] deriving Eq
+newtype SnocPath crumb = SnocPath [crumb] deriving (Eq, Typeable)
 
 instance Monoid (SnocPath crumb) where
    mempty :: SnocPath crumb
@@ -106,6 +114,10 @@ class ExtendPath c crumb | c -> crumb where
   -- | Extend the current 'AbsolutePath' by one crumb.
   (@@) :: c -> crumb -> c
 
+#if __GLASGOW_HASKELL__ >= 708
+deriving instance Typeable ExtendPath
+#endif
+
 -- | A 'SnocPath' from the root.
 type AbsolutePath = SnocPath
 
@@ -116,6 +128,10 @@ type LocalPath = SnocPath
 class ReadPath c crumb | c -> crumb where
   -- | Read the current absolute path.
   absPath :: c -> AbsolutePath crumb
+
+#if __GLASGOW_HASKELL__ >= 708
+deriving instance Typeable ReadPath
+#endif
 
 -- | Lifted version of 'absPath'.
 absPathT :: (ReadPath c crumb, Monad m) => Transform c m a (AbsolutePath crumb)