+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## next
+* Derive `Typeable` instances
+
 ## 2.16.5
 * Allowed building with `base-4.8.0.0`
 * Removed unneeded `Monoid` constraints on `OneT` instances